Support requesting every status explicitly
thor dm:ci:client now accepts a list of different
statuses for the --status or -s flag. This means
that it's possible to either re-run  passed or
failed tasks, or to run modified or skipped jobs
only.

The supported statuses are:

    pass
    fail
    modified
    skipped
    processing

This can also be used to workaround current
synchronization problems. There are no mechanisms
yet that protect us from locking some libraries
to the processing state, or from producing wrong
test results because of misconfiguration on the
client side. This feature allows us to just re-run
faulty reports.