For now, always sync with git before runnings specs
This means that currently every permutation of the
same gem will do a git sync before running specs.
This is because we don't know yet if we already
have a sufficiently recent clone lying around. In
the future, we will probably keep track of the
head SHA1 on the server, so that we know exactly
which revision we're supposed to test.