Only sync master branches when necessary
The server keeps track of library revisions now.
This means that we can check wether our local
clone is synced to the desired revision, and do so
if it isn't.

WARNING: dm:ci:client now syncs the master branches
of all datamapper gems. This means that you should
*never* do any work in master yourself (why would
you anyway?)